+++ demisto_sdk/commands/download/downloader.py
@@ -182,8 +182,9 @@ class Downloader:
         self.num_added_files = 0
         self.init = init
         self.keep_empty_folders = keep_empty_folders
-        if is_sdk_defined_working_offline():
+        if is_sdk_defined_working_offline() and self.run_format:
             self.run_format = False
+            logger.info(f"format is not supported when the DEMISTO_SDK_OFFLINE_ENV environment variable is set.")
 
     def download(self) -> int:
         """